Moved packages to make more sense and added some files to be filled with implementations.
