+++ tests/tensortrade/stochastic/processes/test_heston.py
@@ -1,12 +1,10 @@
 
-import pytest
 
-from tensortrade.data.stochastic import merton
+from tensortrade.stochastic import heston
 
 
-@pytest.mark.skip("Takes too long.")
 def test_shape():
-    frame = merton(
+    frame = heston(
         base_price=7000,
         base_volume=15000,
         start_date='2018-01-01',
