+++ tests/tensortrade/stochastic/processes/test_merton.py
@@ -1,9 +1,12 @@
 
-from tensortrade.data.stochastic import cox
+import pytest
+
+from tensortrade.stochastic import merton
 
 
+@pytest.mark.skip("Takes too long.")
 def test_shape():
-    frame = cox(
+    frame = merton(
         base_price=7000,
         base_volume=15000,
         start_date='2018-01-01',
